fix(BA-4983): merge test_rbac_search.py into test_rbac.py
Merge all search tests from test_rbac_search.py into existing test_rbac.py classes:
- Add status, pagination offset, sorting, and source filter tests to TestRoleSearch
- Add pagination and permission tests to TestScopeManagement
- Add entity search tests with domain_fixture to TestEntityManagement
- Add missing imports: RoleOrder, SearchEntitiesRequest/Response, OrderDirection, RoleOrderField
- Delete test_rbac_search.py as all tests are now in test_rbac.py
